Add FKs to all tables and remove incorrect migration
NULLability is not addressed, that's a separate task.
Remove the RemoveUnusedAwards migration because it was incorrect WRT
the data as it exists in the DB. Resolving that is also another task.
Some users have qualifications create for awards that are not associated with
a program. This is incorrect but apparently some of this information cannot
be lost (e.g. Joe took a 4 star class). However we don't have a training history
for Joe that say's he's taken the 4 star class because history. Resolving
all of that is another task.